requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
requests
python-dotenv
tabulate
joblib
scipy
notebook
